.history/requirements_20210924113702.txt
asgiref==3.4.1
astroid==2.8.0
backcall==0.2.0
black==21.9b0
click==8.0.1
decorator==5.1.0
Django==3.2.7
djangorestframework==3.12.4
ipdb==0.13.9
ipython==7.27.0
isort==5.9.3
jedi==0.18.0
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.3.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
pylint==2.11.1
pytz==2021.1
regex==2021.8.28
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
traitlets==5.1.0
typing-extensions==3.10.0.2
wcwidth==0.2.5
wrapt==1.12.1
